made changes to Animal
Animals live, eat, and die. So do fish and birds, but fish do not fly and birds do not swim. They are both animals so you make them in different tabs under animal so you don't need to copy all the script for each one to live, eat, and die. You only need to script them for their differences. So the Bird Flies, the Fish Swims, and they both Live, Eat, and Die without having to repeat all the codefor each of them. Which makes it clean and also saves time.